requirements.txt
urllib3==1.26.4
pandas-1.2.4
beautifulsoup4-4.9.3
bs4-0.0.1
soupsieve-2.2.1

